third_party/blink/web_tests/external/wpt/css/css-text-decor/text-underline-offset-initial-expected.txt

This is a testharness.js-based test.
FAIL Initial value of text-underline-offset assert_equals: Must be set to value auto as initial value. expected (string) "auto" but got (undefined) undefined
Harness: the test ran to completion.
